+++ app/src/main/java/org/android/go/sopt/data/entity/Follower.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.domain.model
+package org.android.go.sopt.data.entity
 
 
 data class Follower(
