+++ app/src/main/java/org/android/go/sopt/data/entity/MultiViewItem.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.domain.model
+package org.android.go.sopt.data.entity
 
 import org.android.go.sopt.util.type.MultiViewType
 import org.android.go.sopt.util.type.MultiViewType.*
