+++ app/src/main/java/org/android/go/sopt/data/model/remote/request/RequestPostSignUpDto.kt
@@ -1,10 +1,14 @@
-package org.android.go.sopt.data.entity.remote.response
+package org.android.go.sopt.data.model.remote.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponsePostSignUpDto(
+data class RequestPostSignUpDto(
+    @SerialName("id")
+    val id: String,
+    @SerialName("password")
+    val password: String,
     @SerialName("name")
     val name: String,
     @SerialName("skill")
