+++ app/src/main/java/org/android/go/sopt/data/model/remote/response/ResponsePostLoginDto.kt
@@ -1,15 +1,12 @@
-package org.android.go.sopt.data.entity.remote.request
+package org.android.go.sopt.data.model.remote.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import org.android.go.sopt.domain.model.User
 
 @Serializable
-data class RequestPostSignUpDto(
+data class ResponsePostLoginDto(
     @SerialName("id")
     val id: String,
-    @SerialName("password")
-    val password: String,
     @SerialName("name")
     val name: String,
     @SerialName("skill")
