+++ app/src/main/java/org/android/go/sopt/data/service/FollowerService.kt
@@ -1,7 +1,6 @@
 package org.android.go.sopt.data.service
 
-import org.android.go.sopt.data.entity.remote.response.ResponseGetFollowerListDto
-import retrofit2.Call
+import org.android.go.sopt.data.model.remote.response.ResponseGetFollowerListDto
 import retrofit2.http.GET
 import retrofit2.http.Query
