+++ app/src/main/java/org/android/go/sopt/presentation/main/home/HomeViewModel.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.ViewModel
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
 import org.android.go.sopt.GoSoptApplication
-import org.android.go.sopt.data.entity.local.MockRepoDto
-import org.android.go.sopt.domain.model.MultiViewItem.*
+import org.android.go.sopt.data.model.local.MockRepoDto
+import org.android.go.sopt.data.entity.MultiViewItem.*
 import org.android.go.sopt.util.state.LocalUiState
 import org.android.go.sopt.util.state.LocalUiState.*
 import timber.log.Timber
@@ -49,10 +49,7 @@ class HomeViewModel : ViewModel() {
     }
 
     private fun decodeJsonToDtoList(): List<MockRepoDto> {
-        val mockRepoJson = GoSoptApplication.mockRepoJson
-        Timber.d("MOCK REPO JSON: $mockRepoJson")
-
-        return mockRepoJson?.let { json ->
+        return GoSoptApplication.mockJsonString?.let { json ->
             Json.decodeFromString<List<MockRepoDto>>(json)
         } ?: emptyList()
     }
