+++ app/src/main/java/org/android/go/sopt/presentation/main/home/adapter/MultiViewAdapter.kt
@@ -2,8 +2,8 @@ package org.android.go.sopt.presentation.main.home.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import org.android.go.sopt.domain.model.MultiViewItem
-import org.android.go.sopt.domain.model.MultiViewItem.*
+import org.android.go.sopt.data.entity.MultiViewItem
+import org.android.go.sopt.data.entity.MultiViewItem.*
 import org.android.go.sopt.presentation.main.home.adapter.MultiViewHolder.*
 import org.android.go.sopt.util.type.MultiViewType
