+++ app/src/main/java/org/android/go/sopt/presentation/main/home/adapter/MultiViewHolder.kt
@@ -4,7 +4,7 @@ import androidx.databinding.ViewDataBinding
 import androidx.recyclerview.widget.RecyclerView
 import org.android.go.sopt.databinding.ItemHeaderBinding
 import org.android.go.sopt.databinding.ItemRepoBinding
-import org.android.go.sopt.domain.model.MultiViewItem.*
+import org.android.go.sopt.data.entity.MultiViewItem.*
 
 /** sealed class는 추상 클래스로 자식 클래스의 종류를 제한할 수 있다. */
 sealed class MultiViewHolder(
