+++ app/src/main/res/layout/item_repo.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="repo"
-            type="org.android.go.sopt.domain.model.MultiViewItem.Repo" />
+            type="org.android.go.sopt.data.entity.MultiViewItem.Repo" />
 
     </data>
